requirements.txt
fastapi==0.92.0
uvicorn==0.20.0
masonite-orm==2.18.6
psycopg2-binary==2.9.5
strawberry-graphql[fastapi]==0.158.0
pytest==7.2.1
httpx==0.23.3
